+++ task_master.go
@@ -134,7 +134,7 @@ type TaskMaster struct {
 		NewTimer(timer.Setter) timer.Timer
 	}
 	K8sService interface {
-		Client() (k8s.Client, error)
+		Client(string) (k8s.Client, error)
 	}
 	LogService LogService